+++ zarrs/src/array/codec/array_to_bytes/packbits.rs
@@ -15,7 +15,9 @@
 //! ```rust
 //! # let JSON = r#"
 //! {
-//!     "padding_encoding": "start_byte"
+//!     "padding_encoding": "first_byte",
+//!     "first_bit": null,
+//!     "last_bit": null
 //! }
 //! # "#;
 //! # use zarrs_metadata::codec::packbits::PackBitsCodecConfiguration;
@@ -59,10 +61,81 @@ pub(crate) fn create_codec_packbits(metadata: &MetadataV3) -> Result<Codec, Plug
     Ok(Codec::ArrayToBytes(codec))
 }
 
-fn element_size_bits(data_type: &DataType) -> Result<u8, CodecError> {
+struct DataTypeExtensionPackBitsCodecComponents {
+    pub component_size_bits: u64,
+    pub num_components: u64,
+    pub sign_extension: bool,
+}
+
+fn pack_bits_components(
+    data_type: &DataType,
+) -> Result<DataTypeExtensionPackBitsCodecComponents, CodecError> {
     match data_type {
-        DataType::Bool => Ok(1),
-        DataType::Extension(ext) => Ok(ext.codec_packbits()?.size_bits()),
+        DataType::Bool => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 1,
+            num_components: 1,
+            sign_extension: false,
+        }),
+        DataType::UInt8 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 8,
+            num_components: 1,
+            sign_extension: false,
+        }),
+        DataType::Int8 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 8,
+            num_components: 1,
+            sign_extension: true,
+        }),
+        DataType::UInt16 | DataType::Float16 | DataType::BFloat16 => {
+            Ok(DataTypeExtensionPackBitsCodecComponents {
+                component_size_bits: 16,
+                num_components: 1,
+                sign_extension: false,
+            })
+        }
+        DataType::Int16 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 16,
+            num_components: 1,
+            sign_extension: true,
+        }),
+        DataType::UInt32 | DataType::Float32 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 32,
+            num_components: 1,
+            sign_extension: false,
+        }),
+        DataType::Int32 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 32,
+            num_components: 1,
+            sign_extension: true,
+        }),
+        DataType::UInt64 | DataType::Float64 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 64,
+            num_components: 1,
+            sign_extension: false,
+        }),
+        DataType::Int64 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 64,
+            num_components: 1,
+            sign_extension: true,
+        }),
+        DataType::Complex64 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 32,
+            num_components: 2,
+            sign_extension: false,
+        }),
+        DataType::Complex128 => Ok(DataTypeExtensionPackBitsCodecComponents {
+            component_size_bits: 64,
+            num_components: 2,
+            sign_extension: false,
+        }),
+        DataType::Extension(ext) => {
+            let packbits = ext.codec_packbits()?;
+            Ok(DataTypeExtensionPackBitsCodecComponents {
+                component_size_bits: packbits.component_size_bits(),
+                num_components: packbits.num_components(),
+                sign_extension: packbits.sign_extension(),
+            })
+        }
         _ => Err(CodecError::UnsupportedDataType(
             data_type.clone(),
             PACKBITS.to_string(),
@@ -70,16 +143,11 @@ fn element_size_bits(data_type: &DataType) -> Result<u8, CodecError> {
     }
 }
 
-fn elements_size_bytes(data_type: &DataType, num_elements: u64) -> Result<u64, CodecError> {
-    let element_size_bits = element_size_bits(data_type)?;
-    Ok((num_elements * u64::from(element_size_bits)).div_ceil(8))
-}
-
-fn div_rem_8bit(bit: usize, element_size_bits: usize) -> (usize, usize) {
+fn div_rem_8bit(bit: u64, element_size_bits: u64) -> (u64, u8) {
     let (element, element_bit) = bit.div_rem(&element_size_bits);
     let element_size_bits_padded = 8 * element_size_bits.div_ceil(8);
     let byte = (element * element_size_bits_padded + element_bit) / 8;
-    let byte_bit = element_bit % 8;
+    let byte_bit = (element_bit % 8) as u8;
     (byte, byte_bit)
 }
 
@@ -93,7 +161,7 @@ mod tests {
 
     use crate::{
         array::{
-            codec::{ArrayToBytesCodecTraits, CodecOptions},
+            codec::{ArrayToBytesCodecTraits, BytesCodec, CodecOptions},
             element::{Element, ElementOwned},
             ChunkRepresentation,
         },
@@ -126,13 +194,13 @@ mod tests {
     }
 
     #[test]
-    fn codec_packbits() -> Result<(), Box<dyn std::error::Error>> {
+    fn codec_packbits_bool() -> Result<(), Box<dyn std::error::Error>> {
         for encoding in [
             PackBitsPaddingEncoding::None,
-            PackBitsPaddingEncoding::StartByte,
-            PackBitsPaddingEncoding::EndByte,
+            PackBitsPaddingEncoding::FirstByte,
+            PackBitsPaddingEncoding::LastByte,
         ] {
-            let codec = Arc::new(super::PackBitsCodec::new(encoding));
+            let codec = Arc::new(super::PackBitsCodec::new(encoding, None, None).unwrap());
             let data_type = DataType::Bool;
             let fill_value = FillValue::from(false);
 
@@ -153,7 +221,7 @@ mod tests {
                 &chunk_representation,
                 &CodecOptions::default(),
             )?;
-            assert!((encoded.len() as u64) <= 40.div_ceil(&8) + 1);
+            assert!((encoded.len() as u64) <= (40 * 1).div_ceil(&8) + 1);
 
             // Decoding
             let decoded = codec
@@ -185,6 +253,103 @@ mod tests {
             let answer: Vec<bool> =
                 vec![true, false, false, false, true, false, false, false, true];
             assert_eq!(answer, decoded_partial_chunk);
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn codec_packbits_float32() -> Result<(), Box<dyn std::error::Error>> {
+        for encoding in [
+            PackBitsPaddingEncoding::None,
+            PackBitsPaddingEncoding::FirstByte,
+            PackBitsPaddingEncoding::LastByte,
+        ] {
+            let codec = Arc::new(super::PackBitsCodec::new(encoding, None, None).unwrap());
+            let data_type = DataType::Float32;
+            let fill_value = FillValue::from(0.0f32);
+
+            let chunk_shape = vec![NonZeroU64::new(8).unwrap(), NonZeroU64::new(5).unwrap()];
+            let chunk_representation =
+                ChunkRepresentation::new(chunk_shape, data_type.clone(), fill_value).unwrap();
+            let elements: Vec<f32> = (0..40).map(|i| i as f32).collect();
+            let bytes = f32::into_array_bytes(&data_type, &elements)?.into_owned();
+
+            // Encoding
+            let encoded = codec.encode(
+                bytes.clone(),
+                &chunk_representation,
+                &CodecOptions::default(),
+            )?;
+            assert!((encoded.len() as u64) <= (40 * 32).div_ceil(&8) + 1);
+
+            // Decoding
+            let decoded = codec
+                .decode(
+                    encoded.clone(),
+                    &chunk_representation,
+                    &CodecOptions::default(),
+                )
+                .unwrap();
+            assert_eq!(bytes, decoded);
+
+            // Check it matches little endian bytes
+            let decoded = BytesCodec::little()
+                .decode(
+                    encoded.clone(),
+                    &chunk_representation,
+                    &CodecOptions::default(),
+                )
+                .unwrap();
+            assert_eq!(bytes, decoded);
+        }
+        Ok(())
+    }
+
+    #[test]
+    fn codec_packbits_int16() -> Result<(), Box<dyn std::error::Error>> {
+        for last_bit in 11..15 {
+            for first_bit in 0..4 {
+                for encoding in [
+                    PackBitsPaddingEncoding::None,
+                    PackBitsPaddingEncoding::FirstByte,
+                    PackBitsPaddingEncoding::LastByte,
+                ] {
+                    let codec = Arc::new(
+                        super::PackBitsCodec::new(encoding, Some(first_bit), Some(last_bit))
+                            .unwrap(),
+                    );
+                    let data_type = DataType::Int16;
+                    let fill_value = FillValue::from(0i16);
+
+                    let chunk_shape =
+                        vec![NonZeroU64::new(8).unwrap(), NonZeroU64::new(5).unwrap()];
+                    let chunk_representation =
+                        ChunkRepresentation::new(chunk_shape, data_type.clone(), fill_value)
+                            .unwrap();
+                    let elements: Vec<i16> = (-20..20).map(|i| (i as i16) << first_bit).collect();
+                    let bytes = i16::into_array_bytes(&data_type, &elements)?.into_owned();
+
+                    // Encoding
+                    let encoded = codec.encode(
+                        bytes.clone(),
+                        &chunk_representation,
+                        &CodecOptions::default(),
+                    )?;
+                    assert!(
+                        (encoded.len() as u64) <= (40 * (last_bit - first_bit + 1)).div_ceil(8) + 1
+                    );
+
+                    // Decoding
+                    let decoded = codec
+                        .decode(
+                            encoded.clone(),
+                            &chunk_representation,
+                            &CodecOptions::default(),
+                        )
+                        .unwrap();
+                    assert_eq!(elements, i16::from_array_bytes(&data_type, decoded)?);
+                }
+            }
         }
         Ok(())
     }